add claims to jwt

+++ resources/item.py
@@ -1,5 +1,5 @@
 from flask_restful import Resource, reqparse
-from flask_jwt_extended import jwt_required
+from flask_jwt_extended import jwt_required, get_jwt_claims
 from models.item import ItemModel
 
 
@@ -52,7 +52,9 @@ class Item(Resource):
 
     @jwt_required
     def delete(self, name):
-
+        claims = get_jwt_claims()
+        if not claims["is_admin"]:
+            return {"message": "Admin privilege is required"}, 401
         item = ItemModel.find_by_name(name)
         if item:
             item.delete_from_db()
